refactor: Restructure CLI and diagnosis modules for clarity
Split monolithic cli.py into cli/ package with per-command modules.
Rename boot_patterns.py to diagnosis.py, merge patterns/ and fixes/
into diagnose_rules/, extract fix_catalog.py, and remove main.py.
Add stabilization polling to diagnose for running VMs.
