+++ gce_rescue_v2/cli/__main__.py
@@ -1,0 +1,6 @@
+"""Support for python -m gce_rescue_v2.cli."""
+
+import sys
+from . import main
+
+sys.exit(main())
