+++ gce_rescue_v2/cli/preflight.py
@@ -1,0 +1,236 @@
+"""Pre-flight validation helpers for CLI commands."""
+
+import sys
+from typing import Optional
+from ..core.config import VERSION
+
+
+def _create_tracked_client(compute, tracking_label: str):
+    """Create a compute client with a tracking User-Agent header.
+
+    Args:
+        compute: Base compute client (used to extract credentials)
+        tracking_label: Label appended to User-Agent (e.g., 'diagnose-vm-state')
+
+    Returns:
+        Compute API client with User-Agent: gce-rescue-{VERSION}-{tracking_label}
+    """
+    try:
+        from googleapiclient import discovery
+        import googleapiclient.http
+        import google_auth_httplib2
+        import httplib2
+
+        # Verify compute client has real credentials (not a test mock)
+        if not isinstance(getattr(compute, '_http', None), google_auth_httplib2.AuthorizedHttp):
+            return compute
+
+        credentials = compute._http.credentials
+        user_agent = f'gce-rescue-{VERSION}-{tracking_label}'
+
+        def _request_builder(http, *args, **kwargs):
+            headers = kwargs.setdefault('headers', {})
+            headers['user-agent'] = user_agent
+            auth_http = google_auth_httplib2.AuthorizedHttp(
+                credentials, http=httplib2.Http()
+            )
+            return googleapiclient.http.HttpRequest(auth_http, *args, **kwargs)
+
+        return discovery.build(
+            'compute', 'v1', credentials=credentials,
+            cache_discovery=False, requestBuilder=_request_builder
+        )
+    except Exception:
+        return compute
+
+
+def get_gcloud_config(key: str) -> Optional[str]:
+    """
+    Read configuration from gcloud config.
+
+    Args:
+        key: Config key (e.g., 'core/project', 'compute/zone')
+
+    Returns:
+        Config value or None
+    """
+    try:
+        # Try to read from gcloud config
+        import subprocess
+        import platform
+
+        # On Windows, gcloud is a batch file, need shell=True
+        use_shell = platform.system() == 'Windows'
+        result = subprocess.run(
+            ['gcloud', 'config', 'get-value', key],
+            capture_output=True,
+            text=True,
+            timeout=5,
+            shell=use_shell
+        )
+        value = result.stdout.strip()
+        return value if value and value != '(unset)' else None
+    except (subprocess.SubprocessError, FileNotFoundError, OSError):
+        # gcloud not available or error
+        return None
+
+
+def _parse_api_error(e: Exception, vm_name: str, zone: str, project: str = None) -> str:
+    """Parse GCP API error and return user-friendly message."""
+    error_str = str(e)
+
+    if 'was not found' in error_str or 'notFound' in error_str:
+        list_cmd = (f"gcloud compute instances list --project={project}"
+                    if project else "gcloud compute instances list")
+        lines = [
+            f"Instance [{vm_name}] not found.",
+            f"  Zone: {zone}",
+        ]
+        if project:
+            lines.append(f"  Project: {project}")
+        lines.append("")
+        lines.append("To see available instances, run:")
+        lines.append(f"  $ {list_cmd}")
+        lines.append("")
+        return "\n".join(lines)
+
+    if 'Unknown zone' in error_str or (
+        "Invalid value for field" in error_str and "'zone'" in error_str
+    ):
+        lines = [
+            f"Invalid zone '{zone}'.",
+            "",
+            "To see available zones, run:",
+            "  $ gcloud compute zones list",
+            ""
+        ]
+        return "\n".join(lines)
+
+    if 'forbidden' in error_str.lower() or 'permission' in error_str.lower() or '403' in error_str:
+        lines = [
+            "Permission denied.",
+        ]
+        if project:
+            lines.append(f"  Project: {project}")
+        lines.append("")
+        lines.append("To verify project access, run:")
+        lines.append("  $ gcloud projects list")
+        lines.append("")
+        return "\n".join(lines)
+
+    if 'Invalid value for field' in error_str:
+        lines = [
+            "Invalid request parameters.",
+            "",
+            "Verify the following are correct:",
+            f"  Instance: {vm_name}",
+            f"  Zone: {zone}",
+        ]
+        if project:
+            lines.append(f"  Project: {project}")
+        lines.append("")
+        return "\n".join(lines)
+
+    # Fallback: return simplified error
+    return f"API error: {error_str[:200]}\n"
+
+
+def _validate_vm_exists(compute, project: str, zone: str, vm_name: str) -> tuple:
+    """
+    Validate VM exists and is in a valid state for rescue.
+
+    Returns:
+        (success: bool, vm_info: dict or None, error_message: str or None)
+    """
+    try:
+        tracked = _create_tracked_client(compute, 'rescue-vm-preflight')
+        vm = tracked.instances().get(
+            project=project,
+            zone=zone,
+            instance=vm_name
+        ).execute()
+
+        # Check if already in rescue mode
+        metadata = vm.get('metadata', {}).get('items', [])
+        for item in metadata:
+            if item.get('key') == 'rescue-mode':
+                lines = [
+                    f"Instance [{vm_name}] is already in rescue mode.",
+                    "",
+                    "To exit rescue mode and restore the VM, run:",
+                    f"  $ gce-rescue-v2 restore {vm_name} --zone={zone} --project={project}",
+                    ""
+                ]
+                return (False, None, "\n".join(lines))
+
+        # Check VM state
+        status = vm.get('status', 'UNKNOWN')
+        invalid_states = ['STAGING', 'PROVISIONING', 'SUSPENDING', 'SUSPENDED', 'REPAIRING']
+        if status in invalid_states:
+            lines = [
+                f"Instance [{vm_name}] is in state '{status}'.",
+                "",
+                "The VM must be in RUNNING or TERMINATED state to rescue.",
+                "",
+                "To check the current VM status, run:",
+                f"  $ gcloud compute instances describe {vm_name} --zone={zone}"
+                f" --project={project} --format='value(status)'",
+                ""
+            ]
+            return (False, None, "\n".join(lines))
+
+        return (True, vm, None)
+
+    except Exception as e:
+        return (False, None, _parse_api_error(e, vm_name, zone, project))
+
+
+def _check_local_ssds(vm_info: dict) -> list:
+    """Check if VM has Local SSDs attached. Returns list of Local SSD names."""
+    if not vm_info:
+        return []
+
+    local_ssds = []
+    for disk in vm_info.get('disks', []):
+        if disk.get('type') == 'SCRATCH':
+            local_ssds.append(disk.get('deviceName', 'unknown'))
+    return local_ssds
+
+
+def _validate_vm_for_restore(compute, project: str, zone: str, vm_name: str) -> tuple:
+    """
+    Validate VM exists and is in rescue mode for restore.
+
+    Returns:
+        (success: bool, vm_info: dict or None, error_message: str or None)
+    """
+    try:
+        tracked = _create_tracked_client(compute, 'restore-vm-preflight')
+        vm = tracked.instances().get(
+            project=project,
+            zone=zone,
+            instance=vm_name
+        ).execute()
+
+        # Check if in rescue mode
+        metadata = vm.get('metadata', {}).get('items', [])
+        in_rescue_mode = False
+        for item in metadata:
+            if item.get('key') == 'rescue-mode':
+                in_rescue_mode = True
+                break
+
+        if not in_rescue_mode:
+            lines = [
+                f"Instance [{vm_name}] is not in rescue mode.",
+                "",
+                "To put the VM into rescue mode first, run:",
+                f"  $ gce-rescue-v2 rescue {vm_name} --zone={zone} --project={project}",
+                ""
+            ]
+            return (False, None, "\n".join(lines))
+
+        return (True, vm, None)
+
+    except Exception as e:
+        return (False, None, _parse_api_error(e, vm_name, zone, project))
