+++ gce_rescue_v2/core/diagnose_rules/fstab.yaml
@@ -1,5 +1,6 @@
 category: fstab
 fix_guidance: "sudo nano /mnt/sysroot/etc/fstab"
+auto_repair: true
 
 patterns:
   - name: fstab_uuid_not_found
@@ -11,19 +12,17 @@ patterns:
       - 'Timed out waiting for device.*UUID='
       - 'Expecting device.*/dev/disk/by-uuid/'
     fixes:
-      - "Check /mnt/sysroot/etc/fstab for invalid UUID entries"
-      - "Run 'blkid' to see available UUIDs"
       - "Comment out or fix the invalid UUID entry"
 
   - name: fstab_device_timeout
     severity: critical
     description: "Device referenced in /etc/fstab timed out waiting"
     regex:
-      - 'Timed out waiting for device.*/dev/'
+      - 'Timed out waiting for device.*/dev/\S+'
+      - 'Timed out waiting for device dev-\w+'
     fixes:
-      - "Check /mnt/sysroot/etc/fstab for missing or invalid devices"
+      - "Comment out or fix the invalid device entry"
       - "Verify disk is attached: gcloud compute instances describe VM_NAME --zone=ZONE"
-      - "Comment out or fix the invalid entry in /etc/fstab"
 
   - name: fstab_device_not_exist
     severity: critical
@@ -31,11 +30,10 @@ patterns:
     regex:
       - 'special device .* does not exist'
       - 'mount: .*: special device .* does not exist'
-      - 'Device .* does not exist'
+      - 'Device /dev/\S+ does not exist'
     fixes:
-      - "Check /mnt/sysroot/etc/fstab for invalid device paths"
-      - "Verify disk attachments in GCP Console"
       - "Comment out or fix the invalid device entry"
+      - "Verify disk is attached: gcloud compute instances describe VM_NAME --zone=ZONE"
 
   - name: fstab_mount_failed
     severity: critical
@@ -46,9 +44,8 @@ patterns:
       - 'mount.*failed'
       - 'systemd.*Failed to mount'
     fixes:
-      - "Check /mnt/sysroot/etc/fstab for mount errors"
+      - "Comment out or fix the problematic entry"
       - "Verify filesystem type and mount options are correct"
-      - "Check serial console for specific error messages"
 
   - name: fstab_dependency_failed
     severity: critical
@@ -57,8 +54,6 @@ patterns:
       - 'Dependency failed for /[a-zA-Z]'
       - 'Dependency failed for File System Check'
     fixes:
-      - "Check /mnt/sysroot/etc/fstab for entries referencing missing devices"
-      - "Run 'blkid' to list available devices and UUIDs"
       - "Comment out or fix the problematic entry"
 
   - name: fstab_emergency_mode
@@ -70,7 +65,6 @@ patterns:
       - 'Welcome to emergency mode'
       - 'Entering emergency mode'
     fixes:
-      - "Check /mnt/sysroot/etc/fstab for errors"
       - "Review serial console for preceding error messages"
 
   - name: fstab_fsck_failed
@@ -82,6 +76,5 @@ patterns:
       - 'FILE SYSTEM CHECK FAILED'
       - 'UNEXPECTED INCONSISTENCY'
     fixes:
-      - "Run filesystem check: fsck -y /dev/sdb1 (replace with actual device)"
-      - "May need to run: e2fsck -y /dev/sdb1 for ext4 filesystems"
+      - "Run filesystem check: fsck -y /dev/sdX (replace with actual device)"
       - "If persistent, check disk health in GCP Console"
