+++ gce_rescue_v2/tests/test_pattern_loader.py
@@ -4,17 +4,15 @@ import pytest
 import tempfile
 from pathlib import Path
 
-from gce_rescue_v2.core.boot_patterns import (
+from gce_rescue_v2.core.diagnosis import (
     _load_patterns_from_yaml,
     _validate_pattern_file,
     BootErrorPattern,
     BOOT_ERROR_PATTERNS,
-    CATEGORY_FIX_GUIDANCE,
 )
 
 VALID_YAML = """\
 category: test_category
-fix_guidance: "Test fix guidance"
 
 patterns:
   - name: test_pattern
@@ -22,14 +20,10 @@ patterns:
     description: "Test pattern description"
     regex:
       - 'test regex .*'
-    fixes:
-      - "Test fix 1"
-      - "Test fix 2"
 """
 
 MULTI_PATTERN_YAML = """\
 category: multi
-fix_guidance: "Multi fix guidance"
 
 patterns:
   - name: multi_first
@@ -37,16 +31,12 @@ patterns:
     description: "First pattern"
     regex:
       - 'first.*match'
-    fixes:
-      - "Fix first"
   - name: multi_second
     severity: warning
     description: "Second pattern"
     regex:
       - 'second.*match'
       - 'alternate.*second'
-    fixes:
-      - "Fix second"
 """
 
 
@@ -62,7 +52,7 @@ class TestLoadPatternsFromYaml:
 
     def test_valid_yaml_produces_correct_patterns(self, tmp_path):
         _write_yaml(tmp_path, 'test.yaml', VALID_YAML)
-        patterns, fix_guidance = _load_patterns_from_yaml(tmp_path)
+        patterns = _load_patterns_from_yaml(tmp_path)
 
         assert len(patterns) == 1
         assert isinstance(patterns[0], BootErrorPattern)
@@ -71,17 +61,10 @@ class TestLoadPatternsFromYaml:
         assert patterns[0].severity == 'critical'
         assert patterns[0].description == 'Test pattern description'
         assert patterns[0].patterns == ['test regex .*']
-        assert patterns[0].suggested_fixes == ['Test fix 1', 'Test fix 2']
-
-    def test_fix_guidance_loaded(self, tmp_path):
-        _write_yaml(tmp_path, 'test.yaml', VALID_YAML)
-        _, fix_guidance = _load_patterns_from_yaml(tmp_path)
-
-        assert fix_guidance == {'test_category': 'Test fix guidance'}
 
     def test_multiple_patterns_in_one_file(self, tmp_path):
         _write_yaml(tmp_path, 'multi.yaml', MULTI_PATTERN_YAML)
-        patterns, _ = _load_patterns_from_yaml(tmp_path)
+        patterns = _load_patterns_from_yaml(tmp_path)
 
         assert len(patterns) == 2
         assert patterns[0].name == 'multi_first'
@@ -93,46 +76,59 @@ class TestLoadPatternsFromYaml:
     def test_multiple_yaml_files(self, tmp_path):
         _write_yaml(tmp_path, 'a_first.yaml', VALID_YAML)
         _write_yaml(tmp_path, 'b_second.yaml', MULTI_PATTERN_YAML)
-        patterns, fix_guidance = _load_patterns_from_yaml(tmp_path)
+        patterns = _load_patterns_from_yaml(tmp_path)
 
         assert len(patterns) == 3
-        assert len(fix_guidance) == 2
-        assert 'test_category' in fix_guidance
-        assert 'multi' in fix_guidance
 
     def test_empty_directory_raises_error(self, tmp_path):
         with pytest.raises(ValueError, match="No pattern YAML files found"):
             _load_patterns_from_yaml(tmp_path)
+
+    def test_inline_fixes_loaded(self, tmp_path):
+        yaml_with_fixes = """\
+category: test_cat
+patterns:
+  - name: test_fix
+    severity: critical
+    description: "Test with fixes"
+    regex:
+      - 'some error'
+    fixes:
+      - "Fix suggestion 1"
+      - "Fix suggestion 2"
+"""
+        _write_yaml(tmp_path, 'test.yaml', yaml_with_fixes)
+        patterns = _load_patterns_from_yaml(tmp_path)
+        assert patterns[0].fixes == ["Fix suggestion 1", "Fix suggestion 2"]
+
+    def test_missing_fixes_defaults_to_empty(self, tmp_path):
+        _write_yaml(tmp_path, 'test.yaml', VALID_YAML)
+        patterns = _load_patterns_from_yaml(tmp_path)
+        assert patterns[0].fixes == []
 
 
 class TestValidatePatternFile:
     """Tests for _validate_pattern_file."""
 
     def test_missing_category(self):
-        data = {'fix_guidance': 'x', 'patterns': []}
+        data = {'patterns': []}
         with pytest.raises(ValueError, match="missing required field 'category'"):
             _validate_pattern_file(data, 'test.yaml')
 
-    def test_missing_fix_guidance(self):
-        data = {'category': 'x', 'patterns': []}
-        with pytest.raises(ValueError, match="missing required field 'fix_guidance'"):
-            _validate_pattern_file(data, 'test.yaml')
-
     def test_missing_patterns(self):
-        data = {'category': 'x', 'fix_guidance': 'x'}
+        data = {'category': 'x'}
         with pytest.raises(ValueError, match="missing required field 'patterns'"):
             _validate_pattern_file(data, 'test.yaml')
 
     def test_empty_patterns_list(self):
-        data = {'category': 'x', 'fix_guidance': 'x', 'patterns': []}
+        data = {'category': 'x', 'patterns': []}
         with pytest.raises(ValueError, match="'patterns' must be a non-empty list"):
             _validate_pattern_file(data, 'test.yaml')
 
     def test_missing_pattern_field_name(self):
         data = {
             'category': 'x',
-            'fix_guidance': 'x',
-            'patterns': [{'severity': 'critical', 'description': 'd', 'regex': ['r'], 'fixes': ['f']}],
+            'patterns': [{'severity': 'critical', 'description': 'd', 'regex': ['r']}],
         }
         with pytest.raises(ValueError, match="pattern #1 missing required field 'name'"):
             _validate_pattern_file(data, 'test.yaml')
@@ -140,8 +136,7 @@ class TestValidatePatternFile:
     def test_missing_pattern_field_regex(self):
         data = {
             'category': 'x',
-            'fix_guidance': 'x',
-            'patterns': [{'name': 'n', 'severity': 'critical', 'description': 'd', 'fixes': ['f']}],
+            'patterns': [{'name': 'n', 'severity': 'critical', 'description': 'd'}],
         }
         with pytest.raises(ValueError, match="pattern #1 missing required field 'regex'"):
             _validate_pattern_file(data, 'test.yaml')
@@ -149,8 +144,7 @@ class TestValidatePatternFile:
     def test_invalid_severity(self):
         data = {
             'category': 'x',
-            'fix_guidance': 'x',
-            'patterns': [{'name': 'n', 'severity': 'fatal', 'description': 'd', 'regex': ['r'], 'fixes': ['f']}],
+            'patterns': [{'name': 'n', 'severity': 'fatal', 'description': 'd', 'regex': ['r']}],
         }
         with pytest.raises(ValueError, match="invalid severity 'fatal'"):
             _validate_pattern_file(data, 'test.yaml')
@@ -158,8 +152,7 @@ class TestValidatePatternFile:
     def test_invalid_regex_syntax(self):
         data = {
             'category': 'x',
-            'fix_guidance': 'x',
-            'patterns': [{'name': 'n', 'severity': 'critical', 'description': 'd', 'regex': ['[invalid'], 'fixes': ['f']}],
+            'patterns': [{'name': 'n', 'severity': 'critical', 'description': 'd', 'regex': ['[invalid']}],
         }
         with pytest.raises(ValueError, match="invalid regex"):
             _validate_pattern_file(data, 'test.yaml')
@@ -167,8 +160,7 @@ class TestValidatePatternFile:
     def test_empty_regex_list(self):
         data = {
             'category': 'x',
-            'fix_guidance': 'x',
-            'patterns': [{'name': 'n', 'severity': 'critical', 'description': 'd', 'regex': [], 'fixes': ['f']}],
+            'patterns': [{'name': 'n', 'severity': 'critical', 'description': 'd', 'regex': []}],
         }
         with pytest.raises(ValueError, match="must have at least one regex"):
             _validate_pattern_file(data, 'test.yaml')
@@ -177,8 +169,7 @@ class TestValidatePatternFile:
         for severity in ['critical', 'error', 'warning']:
             data = {
                 'category': 'x',
-                'fix_guidance': 'x',
-                'patterns': [{'name': 'n', 'severity': severity, 'description': 'd', 'regex': ['r'], 'fixes': ['f']}],
+                'patterns': [{'name': 'n', 'severity': severity, 'description': 'd', 'regex': ['r']}],
             }
             _validate_pattern_file(data, 'test.yaml')  # Should not raise
 
@@ -189,7 +180,6 @@ class TestInvalidYamlLoading:
     def test_invalid_regex_in_yaml(self, tmp_path):
         bad_yaml = """\
 category: bad
-fix_guidance: "x"
 
 patterns:
   - name: bad_regex
@@ -197,8 +187,6 @@ patterns:
     description: "Bad regex"
     regex:
       - '[invalid'
-    fixes:
-      - "Fix"
 """
         _write_yaml(tmp_path, 'bad.yaml', bad_yaml)
         with pytest.raises(ValueError, match="invalid regex"):
@@ -207,15 +195,12 @@ patterns:
     def test_missing_field_in_yaml(self, tmp_path):
         bad_yaml = """\
 category: bad
-fix_guidance: "x"
 
 patterns:
   - name: no_severity
     description: "Missing severity"
     regex:
       - 'test'
-    fixes:
-      - "Fix"
 """
         _write_yaml(tmp_path, 'bad.yaml', bad_yaml)
         with pytest.raises(ValueError, match="missing required field 'severity'"):
@@ -224,7 +209,6 @@ patterns:
     def test_invalid_severity_in_yaml(self, tmp_path):
         bad_yaml = """\
 category: bad
-fix_guidance: "x"
 
 patterns:
   - name: bad_sev
@@ -232,8 +216,6 @@ patterns:
     description: "Bad severity"
     regex:
       - 'test'
-    fixes:
-      - "Fix"
 """
         _write_yaml(tmp_path, 'bad.yaml', bad_yaml)
         with pytest.raises(ValueError, match="invalid severity 'fatal'"):
@@ -250,15 +232,9 @@ class TestShippedPatterns:
         for pattern in BOOT_ERROR_PATTERNS:
             assert isinstance(pattern, BootErrorPattern)
 
-    def test_fix_guidance_has_entries(self):
-        assert len(CATEGORY_FIX_GUIDANCE) > 0
-
     def test_fstab_patterns_present(self):
         fstab_patterns = [p for p in BOOT_ERROR_PATTERNS if p.category == 'fstab']
         assert len(fstab_patterns) == 7
-
-    def test_fstab_fix_guidance_present(self):
-        assert 'fstab' in CATEGORY_FIX_GUIDANCE
 
     def test_all_patterns_have_valid_severity(self):
         for pattern in BOOT_ERROR_PATTERNS:
@@ -269,3 +245,11 @@ class TestShippedPatterns:
         for pattern in BOOT_ERROR_PATTERNS:
             for regex in pattern.patterns:
                 re.compile(regex)  # Should not raise
+
+    def test_fstab_patterns_have_inline_fixes(self):
+        """All fstab patterns should have inline fixes from merged YAML."""
+        fstab_patterns = [p for p in BOOT_ERROR_PATTERNS if p.category == 'fstab']
+        for pattern in fstab_patterns:
+            assert len(pattern.fixes) > 0, (
+                f"Pattern '{pattern.name}' has no inline fixes"
+            )
