+++ setup.py
@@ -63,8 +63,7 @@ setup(
           'startup_scripts/*.sh',
           'startup_scripts/*.ps1',
           'startup_scripts/fixes/*.sh',
-          'core/patterns/*.yaml',
-          'core/patterns/README.md',
+          'core/diagnose_rules/*.yaml',
       ],
   },
   include_package_data = True,
